+++ docs/changelog.md
@@ -62,6 +62,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: components can be inherited from parent issue
 * feat: labels can be inherited from parent issue
 * feat: description can be a multi-line description
+* feat: assign subtask to current user or inherit from parent issue
 
 [Unreleased]: https://github.com/codescape/jira-multiple-subtasks/compare/22.5.0...HEAD
 [22.5.0]: https://github.com/codescape/jira-multiple-subtasks/tree/22.5.0